+++ core-ng/src/main/java/core/framework/internal/log/filter/BytesLogParam.java
@@ -1,5 +1,7 @@
 package core.framework.internal.log.filter;
 
+import core.framework.util.Strings;
+
 import java.nio.charset.Charset;
 import java.util.Set;
 
@@ -8,7 +10,7 @@ import static java.nio.charset.StandardCharsets.UTF_8;
 /**
  * @author neo
  */
-public class BytesLogParam implements LogParam {
+public class BytesLogParam implements LogParam {    // for text based value, detect if in json format, and mask if needed
     private final byte[] bytes;
     private final Charset charset;
 
@@ -25,12 +27,77 @@ public class BytesLogParam implements LogParam {
     public void append(StringBuilder builder, Set<String> maskedFields, int maxParamLength) {
         if (bytes == null) {
             builder.append("null");
-        } else if (bytes.length > maxParamLength) {
-            var value = new String(bytes, 0, maxParamLength, charset);
-            builder.append(value, 0, value.length() - 1);   // not use last char as can be cut off bytes
-            builder.append("...(truncated)");
-        } else {
-            builder.append(new String(bytes, charset));
+            return;
         }
+
+        boolean truncate = false;
+        String value;
+        if (bytes.length > maxParamLength) {
+            value = new String(bytes, 0, maxParamLength, charset);
+            truncate = true;
+        } else {
+            value = new String(bytes, charset);
+        }
+        if (shouldMask(value, maskedFields)) {
+            builder.append(filter(value, maskedFields));
+        } else {
+            if (truncate) {
+                builder.append(value, 0, value.length() - 1);
+            } else {
+                builder.append(value);
+            }
+        }
+        if (truncate) {
+            builder.append("...(truncated)");
+        }
+    }
+
+    StringBuilder filter(String value, Set<String> maskedFields) {
+        var builder = new StringBuilder(value);
+        for (String maskedField : maskedFields) {
+            int current = -1;
+            while (true) {
+                current = builder.indexOf('\"' + maskedField + '\"', current);
+                if (current < 0) break;
+                int[] range = maskRange(builder, current + maskedField.length() + 2);    // start from last double quote
+                if (range == null) break;
+                builder.replace(range[0], range[1], "******");      // with benchmark, StringBuilder.replace is the fastest way to mask substring
+                current = range[1];
+            }
+        }
+        return builder;
+    }
+
+    // find first json "string" range from start, only mask string field value, ignore if target field value is list or object
+    private int[] maskRange(StringBuilder builder, int start) {
+        boolean escaped = false;
+        int maskStart = -1;
+        int length = builder.length();
+        for (int index = start; index < length; index++) {
+            char ch = builder.charAt(index);
+            if (ch == '\\') {
+                escaped = true;
+            } else if (!escaped && maskStart < 0 && ch == '\"') {
+                maskStart = index + 1;
+            } else if (!escaped && maskStart >= 0 && ch == '\"') {
+                return new int[]{maskStart, index};
+            } else if (maskStart < 0 && (ch == ',' || ch == '{' || ch == '}' || ch == '[' || ch == ']')) {
+                return null;    // not found start double quote, and reached next field
+            } else {
+                escaped = false;
+            }
+        }
+        if (maskStart >= 0) return new int[]{maskStart, length};  // json can be truncated, mask rest if found start quote
+        return null;
+    }
+
+    private boolean shouldMask(String value, Set<String> maskedFields) {
+        // only consider json string starts with '{', ignore other rare cases like array starts with '['
+        if (!Strings.startsWith(value, '{')) return false;
+        for (String maskedField : maskedFields) {
+            int index = value.indexOf('\"' + maskedField + '\"');
+            if (index >= 0) return true;
+        }
+        return false;
     }
 }