+++ core-ng/src/main/java/core/framework/internal/log/filter/BytesMapLogParam.java
@@ -3,28 +3,30 @@ package core.framework.internal.log.filter;
 import java.util.Map;
 import java.util.Set;
 
+import static java.nio.charset.StandardCharsets.UTF_8;
+
 /**
  * @author neo
  */
-public class MapLogParam implements LogParam {
-    private final Map<String, String> values;
+public class BytesMapLogParam implements LogParam {
+    private final Map<String, byte[]> values;
 
-    public MapLogParam(Map<String, String> values) {
+    public BytesMapLogParam(Map<String, byte[]> values) {
         this.values = values;
     }
 
+    // not masking the bytes values, thought it can be in json format
+    // currently sensitive info won't be handled in batch via cache/redis, so not to add unnecessary overhead
     @Override
     public void append(StringBuilder builder, Set<String> maskedFields, int maxParamLength) {
         int maxLength = builder.length() + maxParamLength;
         builder.append('{');
         int index = 0;
-        for (Map.Entry<String, String> entry : values.entrySet()) {
+        for (Map.Entry<String, byte[]> entry : values.entrySet()) {
             if (index > 0) builder.append(", ");
-            String key = entry.getKey();
-            builder.append(key).append('=');
-
-            if (maskedFields.contains(key)) builder.append("******");
-            else builder.append(entry.getValue());
+            builder.append(entry.getKey())
+                    .append('=')
+                    .append(new String(entry.getValue(), UTF_8));
 
             if (builder.length() >= maxLength) {
                 builder.setLength(maxLength);