+++ core-ng/src/main/java/core/framework/internal/web/response/BeanBody.java
@@ -1,13 +1,11 @@
 package core.framework.internal.web.response;
 
-import core.framework.internal.log.filter.JSONLogParam;
+import core.framework.internal.log.filter.BytesLogParam;
 import io.undertow.io.Sender;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import java.nio.ByteBuffer;
-
-import static java.nio.charset.StandardCharsets.UTF_8;
 
 /**
  * @author neo
@@ -23,7 +21,7 @@ public final class BeanBody implements Body {
     @Override
     public void send(Sender sender, ResponseHandlerContext context) {
         byte[] body = context.writer.toJSON(bean);
-        LOGGER.debug("[response] body={}", new JSONLogParam(body, UTF_8));
+        LOGGER.debug("[response] body={}", new BytesLogParam(body));
         sender.send(ByteBuffer.wrap(body));
     }
 }